Moved converter to its appropriate namespace
